Backend/analysismedi/requirements.txt
streamlit
pytesseract
pillow
google-generativeai
